Make temporary files and directories movable

+++ cpp/common/file_util.cc
@@ -3,6 +3,7 @@
 #include <cstdio>
 #include <filesystem>
 #include <fstream>
+#include <iostream>
 #include <random>
 #include <sstream>
 
@@ -42,7 +43,9 @@ TempFile::TempFile(std::string_view prefix) {
 
 TempFile::~TempFile() {
   // Delete the owned temporary file.
-  std::filesystem::remove(path_);
+  if (!path_.empty()) {
+    std::filesystem::remove(path_);
+  }
 }
 
 const std::filesystem::path& TempFile::GetPath() const { return path_; }
@@ -65,7 +68,9 @@ TempDir::TempDir(std::string_view prefix) {
 
 TempDir::~TempDir() {
   // Delete the owned temporary file.
-  std::filesystem::remove_all(path_);
+  if (!path_.empty()) {
+    std::filesystem::remove_all(path_);
+  }
 }
 
 const std::filesystem::path& TempDir::GetPath() const { return path_; }
